fee & payment models

+++ src/main/java/pps/gestorClub_api/entities/FeeEntity.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Entity
 @AllArgsConstructor
@@ -25,5 +25,5 @@ public class FeeEntity {
     private Integer year;
 
     @Column(name = "amount", nullable = false)
-    private Double amount;
+    private BigDecimal amount;
 }
